Add the output of depmod into kernel module .tgz file
CQID: WIND00432340

do_rootfs() runs depmod on the final rootfs, so the depmod files are
in the rootfs image. But the kernel module .tgz doesn't have the files.
If you un-tar the new version of the kernel module .tgz on the board,
it will need to re-run depmod. The avoid this we are running depmod in
kernel_do_deploy() before generating the kernel .tgz.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20141118

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>