sysklog: comment out configuration regarding xconsole
I commented out these configurations for the following reasons:
We don't have xconsole utility in OE. These configurations are useless
for us.

Besides, the /dev/xconsole device is created in the init script which means
if these configuration are enabled, we need to add a script in sysklogd which
creates /dev/xconsole and is invoked in ExecStartPre in the sysklogd's service
file.

I think it's better that we just comment these lines out.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20140921

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>