toolchain-scripts.bbclass: do not expand vars which starts with multilib prefixes
Issue: LIN8-1413

When set MULTILIBS with two or more items, such as

MULTILIBS = "multilib:lib32 multilib:lib64"

It expands depends once in mutlib.bbclass, and expand again in
toolchain-scripts.bbclass that causes error:

| ERROR: Nothing PROVIDES 'virtual/lib32-lib64-libc'.

Check before expand a var. If starts with a multilib prefix, ignore it.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20151113

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>