Use TUNE_CCARGS for target compilation
ISSUE: LIN8-1048, LIN8-1655

The initial gcc configuration is using TARGET_CFLAGS, but this is
probably not what is intended, because TARGET_CFLAGS doesn't include
any architecture-specific flags.

For PPC64, it's possible to end up with gcc configured to pass -a64 to
the assembler, but be itself getting invoked without -m64, which can result
in alignment problems in generated code. Solution: Instead of setting
CFLAGS_FOR_TARGET to TARGET_CFLAGS, set it to TARGET_CFLAGS + TUNE_CCARGS.

However, for crosssdk cases, we need to prevent that from having any
effect; this can be done by emptying out the TUNE_*ARGS flags in the
crosssdk bbclass, because there is no "target tune" in crosssdk cases.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20151119

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>