gcc-sanitizers: allow empty for lsan and tsan packages
Issue: LIN9-1318

gcc checks threads and leak sanitizers support when do configure. If not
supported, packages of tsan and lsan are empty. If add such package to
image it fails to build:

| ERROR: core-image-minimal-1.0-r0 do_rootfs: libtsan-dev not found in the feeds
| ERROR: core-image-minimal-1.0-r0 do_rootfs: This is often caused by an empty
| package declared in a recipe's PACKAGES variable. (Empty packages are not
| constructed unless ALLOW_EMPTY_<pkg> = '1' is used.)

Set ALLOW_EMPTY for these packages to resolve the issue.

(LOCAL REV: NOT UPSTREAM) -- Will send to oe-core on 20161012

Signed-off-by: Kai Kang <kai.kang@windriver.com>