systemd: avoid boot failures for mips64 and mipsel
Issue: LIN9-1849

If we start a systemd based qemumips64/qemumipsel target with 'pam' enabled
in DISTRO_FEATURES, we could not login in successfully. After you provide
username and password to the login prompt, the login prompt just comes up
again. Besides, there are other boot failures.

Append '-O0' flag for mips64/mipsel target as a workaround to this problem.

(LOCAL REV: NOT UPSTREAM) -- sent to oe on 20161009

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Robert Yang <liezhi.yang@windriver.com>