libfm: disable build static library
Issue: LIN9-2198

libfm and fmlib install same name static library libfm.a that causes error:

| ERROR: libfm-1.2.4-r0 do_populate_sysroot: The recipe libfm is trying
| to install files into a shared area when those files already exist.
| Those files and their manifest location are:
| .../wrlinux-std-sato_nxp-t4xxx_world9/tmp/sysroots/nxp-t4xxx/usr/lib64/libfm.a
| Matched in b'manifest-nxp-t4xxx-fmlib.populate_sysroot'

Disable build static library for libfm to solve the issue. And it is
disabled to build static library of libfm in popular linux distributions
either, such as ubuntu and fedora.

(LOCAL REV: NOT UPSTREAM) -- Send to oe-core on 20161117

Signed-off-by: Kai Kang <kai.kang@windriver.com>