gdk-pixbuf: Fix CVE-2016-6352
Issue: LIN9-1719

Backported from:
https://git.gnome.org/browse/gdk-pixbuf/commit/?id=88af50a864195da1a4f7bda5f02539704fbda599

There is some redundancy between the ico directory and the
bitmap image header. If the two disagree on the icon dimensions,
just toss the image, instead of risking crashes or OOM later. Also
add some more debug spew that helped in tracking this down, and
make error messages more unique.

The commit also includes a test image that has an example of
this discrepancy and triggers the early exit.

https://bugzilla.gnome.org/show_bug.cgi?id=769170

(LOCAL REV: NOT UPSTREAM) -- WR Specific

Signed-off-by: Wenzong Fan <wenzong.fan@windriver.com>