icu: avoid host contamination
Issue: LIN9-2199

icu provides cross script icu-config. When compile package which calls
'icu-config --cppflags', it returns '-I/usr/include' which causes host
contamination.

The root cause is when install cross script by binconfig.bbclass, the
pure prefix(/usr) is not replaced. Then the value of $(prefix)/include
points to the host directory. Fix by replace it with $(includedir) which
is updated properly.

(LOCAL REV: NOT UPSTREAM) -- Send to oe-core on 20161122

Signed-off-by: Kai Kang <kai.kang@windriver.com>