bitbake.conf: add VOLATILE_LOG_DIR variable
Issue: LIN9-2098

The default value is "yes" which results in the /var/log being a link
pointing to /var/volatile/log which is on tmpfs.

The user could override this value to "no" which causes /var/log to be
a directory on persistent storage.

[YOCTO #6132]

(LOCAL REV: NOT UPSTREAM) -- sent to oe on 20161122

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>