apr: fix off_t size can't match when configure and in target glibc
Issue: LIN9-2111

When subversion run on x86 and lib32 on x86-64, it use the APR's
apr.h header file. But when configure the APR, APR meets the cross
compiling, it was hardcoded in configure.in in apr source code. As
the following:

	APR_CHECK_SIZEOF_EXTENDED([#include <sys/types.h>], off_t, 8)

It pass 8 bytes to off_t when meets cross compiling, but on x86 or lib32
the off_t in glibc was 4 bytes, so it let the application who use apr.h
go to wrong.

Such as subversion:

	svnadmin create test

So we should let the APR detect the correct off_t when cross compiling,
change it to the following:

	AC_CHECK_SIZEOF(off_t)

(LOCAL REV: NOT UPSTREAM) -- Send to oe-core 20161216

Signed-off-by: Dengke Du <dengke.du@windriver.com>