openssl: CVE-2017-3732
Issue: LIN9-3128

There is a carry propagating bug in the x86_64 Montgomery squaring
procedure. No  EC algorithms are affected. Analysis suggests that attacks against RSA
and DSA  as a result of this defect would be very difficult to perform and are
not believed likely. Attacks against DH are considered just feasible
(although very difficult) because most of the work necessary to deduce information
about a private key may be performed offline. The amount of resources
required for such an attack would be very significant and likely only
accessible to a limited number of attackers. An attacker would
additionally need online access to an unpatched system using the target
private key in a scenario with persistent DH parameters and a private
key that is shared between multiple clients. For example this can occur
by default in OpenSSL DHE based SSL/TLS ciphersuites. Note: This issue is
very similar to CVE-2015-3193 but must be treated as a separate problem

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20170207

Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>