openssl: force soft link to avoid rare race
Issue: LIN10-693

This patch works around a rare parallel build race condition using
the force option when soft linking.

The error seen is:

ln: failed to create symbolic link 'libssl.so': File exists
make[4]: *** [Makefile.shared:171: link_a.gnu] Error 1
make[4]: Leaving directory
'/.../build/tmp-glibc/work/x86_64-linux/openssl-native/1.0.2k-r0/openssl-1.0.2k'

Just add the -f flag to the platform independent soft link code to avoid
the collision.  This is reasonable since this Makefile removes the link target
before creating a new soft link. The Makefile was written this way to support
platforms that don't allow forcing a softlink to overwrite an existing link.
Only builds on Linux are supported so that's not a requirement for oe-core recipes.

The openssl team is rewriting their build files so it's not
appropriate for openssl upstream and fixing the root cause of
the race condition was also not pursued.

(LOCAL REV: NOT UPSTREAM) -- Testing locally before sending to oe-core.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>