tcl: remove host path from tclConfig.sh
Issue: LIN10-1286

The tclConfig.sh is also used by other packages (such as expect) for
cross-compiling, the host path from it can't be removed directly in
the do_install step.

With PACKAGE_PREPROCESS_FUNCS to remove host path and avoid the
crossscripts installed to target.

(LOCAL REV: NOT UPSTREAM) -- Sent to yocto on 20171011

Signed-off-by: Wenzong Fan <wenzong.fan@windriver.com>