wpa-supplicant/hostapd: WPA packet number reuse with replayed messages and key reinstallation
A vulnerability was found in how a number of implementations can be
triggered to reconfigure WPA/WPA2/RSN keys (TK, GTK, or IGTK) by
replaying a specific frame that is used to manage the keys. Such
reinstallation of the encryption key can result in two different types
of vulnerabilities: disabling replay protection and significantly
reducing the security of encryption to the point of allowing frames to
be decrypted or some parts of the keys to be determined by an attacker
depending on which cipher is used.

CVE list:
CVE-2017-13077
CVE-2017-13078
CVE-2017-13079
CVE-2017-13080
CVE-2017-13081
CVE-2017-13082
CVE-2017-13086
CVE-2017-13087
CVE-2017-13088

Issue: LIN9-5553 LIN9-5565 LIN9-5596 LIN9-5557 LIN9-5552
       LIN9-5526 LIN9-5544 LIN9-5542 LIN9-5522

Signed-off-by: Haiqing Bai <Haiqing.Bai@windriver.com>