testsdkext: remove workspace/sources to avoid failure in case of multilib
Issue: LIN1018-603

When multilib is enabled, there are multiple environment scripts, and the
test cases for eSDK are executed for each environment script.

And we will have the following problem when executing test cases for the
second environment script.

  ERROR: Source tree path /.../workspace/sources/librdfa already exists and is not empty

So after executing test cases for one environment, we clean up the sources
diretory to avoid such failure.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20180508

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>