weston: Remove machine specific append
If you try to build a system with multiple BSPs, one of which is qemux86
or qemux86-64, the overall system configuration will change and all of the
graphics packages will end up being rebuilt each time.

For a package based system, the PR values will also be incremented each
time.  The end result will be an ever growing set of PR values as well as
being unable to tell which configured version of the graphics components
are really being deployed.

The solution here was to always include the fbdev backend when an x86 BSP
is enabled.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core list on 20181017

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>