layer.conf: Add LAYERSERIES_COMPAT
Issue: US117235

Besides the 'rocko' in order to be compatible with the core layer, we
also add a 'wrl' tag even this is not a layer owned by Wind River,
but we do claim that the distros and machines in this layer are
supported by Wind River. So add the 'wrl' tag in the
LAYERSERIES_COMPAT to make sure that the distros and machines in this
layer are listed by default in the output of
'--list-{machines,distros}'.

(LOCAL REV: NOT UPSTREAM) -- Wind River Specific

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>