bind: fix CVE-2019-6471
Issue: LIN1019-2883

Fix CVE-2019-6471 for bind.

(LOCAL REV: NOT UPSTREAM) -- I sent one commit to fix both LIN1019-2883
and CVE-2018-5743 to oe-core and it has been merged to master but not
branch zeus. There are 6 patches for CVE-2018-5743 which are too many,
so split the commit to fix LIN1019-2883 first.

Signed-off-by: Kai Kang <kai.kang@windriver.com>