openssl: fix CVE-2020-1967
Issue: LIN1019-4280

Server or client applications that call the SSL_check_chain()
function during or after a TLS 1.3 handshake may crash due to
a NULL pointer dereference as a result of incorrect handling of
the "signature_algorithms_cert" TLS extension.

The crash occurs if an invalid or unrecognised signature
algorithm is received from the peer. This could be exploited
by a malicious peer in a Denial of Service attack.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 2020/04/22

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>