arm64: continue loading even if kaslr-seed is not wiped to zero
Issue: LINUXEXEC-7949

The commit c3f043241a866a (arm64: Add support to supply 'kaslr-seed' to secondary kernel)
add kaslr-seed support. And consider the primary kernel reads the 'kaslr-seed' and
wipes it to 0. But in the situation, 'CONFIG_RANDOMIZE_BASE' was not set to y in the
primary kernel and ATF firmware has set the 'kaslr-seed' dtb property with non-zero, Thus
will return error.

So in the above case, continue loading the segments with no kaslr supported situation.

Upstream-Status: Pending

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>