kernel-yocto.bbclass: chdir to ${WORKDIR} for do_kernel_checkout
Issue: LINCD-5019

It chdirs to ${S} at the beginning of task do_kernel_checkout. Then it
removes ${S} when it still resides in ${S}. It may fail to run the task
do_kernel_checkout when bitbake is called by third-part wrapper script.
So chdir to ${WORKDIR} by default for do_kernel_checkout. And it will
chdir to ${S} afterwards in task do_kernel_checkout.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20210423

Signed-off-by: Kai Kang <kai.kang@windriver.com>