rt-tests: Update rt-tests
Issue: LINCD-4890

When run cyclictest, the parameter enable NUMA. but in some BSP
which not support NUMA, will causes the test program to exit
directly and does not carry out subsequent operations, the
latest changes have fixed this problem. so update to the
latest branch to resolve.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20210513

Signed-off-by: Zqiang <qiang.zhang@windriver.com>