mklibs: remove recipes and class
Issue: LIN1021-852

This is not enabled or tested by default, and has never been
ported to python 3 upstream[1], which means it doesn't work at all
with plain poky. If you need it, please put it in a separate layer
and/or modernize to work with py3.

https://salsa.debian.org/installer-team/mklibs/-/blob/master/src/mklibs

Signed-off-by: Alexander Kanavin <alex.kanavin@gmail.com>
Signed-off-by: Jacob Kroon <jacob.kroon@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>

(master rev: 908df863b419d1cad7317153101fc827e7e3a354)
(LOCAL REV: NOT UPSTREAM) -- sent to hardknott on 20210728

Signed-off-by: Changqing Li <changqing.li@windriver.com>