rt-tests: update SRCREV
Issue: LIN1021-138

The commit '6af88067da16c6e69243b376c3b6454ec694460b' is not available
in the repo any more. I think it's due to some mis-operation of the upstream
repo in the past few hours.

Update the SRCREV to use the latest commit on unstable/devel/latest branch.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core on 20210517

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mingli Yu <mingli.yu@windriver.com>