libgcrypt: fix CVE-2021-33560
Issue: LIN1021-421

CVE-2021-33560:
Libgcrypt before 1.8.8 and 1.9.x before 1.9.3 mishandles ElGamal
encryption because it lacks exponent blinding to address a side-channel
attack against mpi_powm, and the window size is not chosen
appropriately. (There is also an interoperability problem because the
selection of the k integer value does not properly consider the
differences between basic ElGamal encryption and generalized ElGamal
encryption.) This, for example, affects use of ElGamal in OpenPGP.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2021-33560

Patch from:
https://git.gnupg.org/cgi-bin/gitweb.cgi?p=libgcrypt.git;a=commit;h=632d80ef30e13de6926d503aa697f92b5dbfbc5e

(LOCAL REV: NOT UPSTREAM) -- The upstream version 1.9.3 is not affected.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>