kernel: remove recipes for old v5.4 ktypes
We've dumped the v5.4/<yocto-name> branches and all 116 v5.4.x stable tags
and over 12,500 v5.4 backports from stable to reduce our download size.

The v5.4 kernel recipes can be problematic, since AUTOINC ones like -rt
will validate SRCREV at parse time and trigger a parse fail:

 FetchError: Fetcher failure: Unable to resolve 'v5.4/standard/preempt-rt/base'
 in upstream git repository in git ls-remote output for linux-yocto.git

I entertained the idea of pointing KBRANCH to just vanilla v5.4 but
there are other v5.4.x SRCREV in there which now won't exist and leaving
the recipes in place is just asking for another corner case parse error.

Plus, muddling with the KBRANCH and SRCREV in v5.4 recipes sends the
wrong message, indicating they are supported in some kind of fashion,
when they are not.  If they are gone, the message is 100% clear.

So, after thinking about it a bunch, I'm convinced the right thing to
do here is just push the v5.4 recipes off a cliff, which is exactly what
yocto is going to do eventually anyways.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>