assimp: Remove it
Issue: LIN1021-1441

Remove it since it has license issues:

* scripts/StepImporter/schema_ifc2x3.exp
Copyright by: International Alliance for Interoperability, 1996-2005 All rights
reserved. No part of this documentation may be reproduced, stored in a
retrieval system, or transmitted in any form or by any means, without the prior
written permission of the owner

* test/models-nonbsd/X/dwarf-Read-Me.txt
RESTRICTIONS: This model pack is available for use in freeware, shareware,
commercial games/software with the following restrictions:- **You may not
sell/re-sell this model pack or claim it as your own. ***You may not
redistribute this pack in some other model pack through a website or on a
compilation CD of any kind, without my written consent. Psi
http://www.psionic3d.co.uk

* test/models-nonbsd/B3D/turtle.source.txt
Copyright 2004, Psionic Design e-mail: psionic@blueyonder.co.uk Used with
permission. RESTRICTIONS: This model pack is available for use in freeware,
shareware, commercial games/software with the following restrictions:- **You
may not sell/re-sell this model pack or claim it as your own. ***You may not
redistribute this pack in some other model pack through a website or on a
compilation CD of any kind, without my written consent.

And no recipes depend on it, so remove it.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20210907

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>