openssl: upgrade 1.1.1l -> 1.1.1m
Issue: LIN1019-7821

Includes fix for CVE-2021-4160:
There is a carry propagation bug in the MIPS32 and MIPS64 squaring procedure.
Many EC algorithms are affected, including some of the TLS 1.3 default curves.
Impact was not analyzed in detail, because the pre-requisites for attack are
considered unlikely and include reusing private keys. Analysis suggests that
attacks against RSA and DSA as a result of this defect would be very difficult
to perform and are not believed likely. Attacks against DH are considered just
feasible (although very difficult) because most of the work necessary to deduce
information about a private key may be performed offline. The amount of
resources required for such an attack would be significant. However, for an
attack on TLS to be meaningful, the server would have to share the DH private
key among multiple clients, which is no longer an option since CVE-2016-0701.
This issue affects OpenSSL versions 1.0.2, 1.1.1 and 3.0.0. It was addressed in
the releases of 1.1.1m and 3.0.1 on the 15th of December 2021. For the 1.0.2
release it is addressed in git commit 6fc1aaaf3 that is available to premium
support customers only. It will be made available in 1.0.2zc when it is
released. The issue only affects OpenSSL on MIPS platforms. Fixed in OpenSSL
3.0.1 (Affected 3.0.0). Fixed in OpenSSL 1.1.1m (Affected 1.1.1-1.1.1l). Fixed
in OpenSSL 1.0.2zc-dev (Affected 1.0.2-1.0.2zb).

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2021-4160

(LOCAL REV: NOT UPSTREAM) -- wrspecific

[OP: keep old openssl 1.1.1l recipe]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>