wpa-supplicant: Security fixes for CVE-2022-23303 CVE-2022-23304
Issue: LIN1021-2458
Issue: LIN1021-2459
Issue: LIN1021-2646

CVE-2022-23303:
The implementations of SAE in hostapd before 2.10 and wpa_supplicant
before 2.10 are vulnerable to side channel attacks as a result of cache
access patterns. NOTE: this issue exists because of an incomplete fix
for CVE-2019-9494.

CVE-2022-23304:
The implementations of EAP-pwd in hostapd before 2.10 and wpa_supplicant
before 2.10 are vulnerable to side-channel attacks as a result of cache
access patterns. NOTE: this issue exists because of an incomplete fix
for CVE-2019-9495.

References:
https://nvd.nist.gov/vuln/detail/CVE-2022-23303
https://nvd.nist.gov/vuln/detail/CVE-2022-23304

Patches from: https://w1.fi/security/2022-1/

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20220214

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>