connman: Security fixes for CVE-2022-23096 CVE-2022-23097 CVE-2022-23098
Issue: LIN1021-2630
Issue: LIN1021-2631
Issue: LIN1021-2632

CVE-2022-23096:
An issue was discovered in the DNS proxy in Connman through 1.40. The
TCP server reply implementation lacks a check for the presence of
sufficient Header Data, leading to an out-of-bounds read.

CVE-2022-23097:
An issue was discovered in the DNS proxy in Connman through 1.40.
forward_dns_reply mishandles a strnlen call, leading to an out-of-bounds
read.

CVE-2022-23098:
An issue was discovered in the DNS proxy in Connman through 1.40. The
TCP server reply implementation has an infinite loop if no data is
received.

References:
https://nvd.nist.gov/vuln/detail/CVE-2022-23096
https://nvd.nist.gov/vuln/detail/CVE-2022-23097
https://nvd.nist.gov/vuln/detail/CVE-2022-23098

Patches from:
https://git.kernel.org/pub/scm/network/connman/connman.git/commit/?id=e5a313736e13c90d19085e953a26256a198e4950
https://git.kernel.org/pub/scm/network/connman/connman.git/commit/?id=d8708b85c1e8fe25af7803e8a20cf20e7201d8a4

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-core on 20220310

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>