gcc: Fix compile of gcc plugins
Issue: LIN1021-3524
LOCAL REV: NOT UPSTREAM - rev: 001

Without this, compiling gcc plugins will fail with an error similar to:
[...]
fatal error: ./config/i386/linux64.h: No such file or directory

In Yocto, we set up compiling gcc-cross out of tree. Which in turn makes the generated headers end up in B. The tm.h header will include generated headers that are expected in plugin/include/config/*.

For example, the linux64.h header, when generating gcc-cross for x86-64, will end up in tm.h header as:
     include "./config/i386/linux64.h"
On the other hand, the make rule `install-plugin` in gcc/Makefile.in will install the linux64.h assuming that it is generated in the sources directory and because this is not the case in our setup, the Makefile ends up installing it in plugin_includedir/`basename $$path` which ends up installing the header in [..]plugin/include as opposed to   [..]plugin/include/config/i386 (as expected by the generator of tm.h).

The included patch modifies the Makefile rule to match the assumption of gcc-cross being compiled out of tree.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>

This is a backport of an upstream oe-core patches:
92167f8e02 gcc: Fix compile of gcc plugins
5f0f00f098 gcc: Fix typo in linux64.h install patch

Signed-off-by: Sundeep KOKKONDA <sundeep.kokkonda@windriver.com>