openssl: upgrade 1.1.1o -> 1.1.1p
Issue: LIN1019-8474

In addition to the c_rehash shell command injection identified in
CVE-2022-1292, further circumstances where the c_rehash script does not
properly sanitise shell metacharacters to prevent command injection were found
by code review. When the CVE-2022-1292 was fixed it was not discovered that
there are other places in the script where the file names of certificates being
hashed were possibly passed to a command executed through the shell. This
script is distributed by some operating systems in a manner where it is
automatically executed. On such operating systems, an attacker could execute
arbitrary commands with the privileges of the script. Use of the c_rehash
script is considered obsolete and should be replaced by the OpenSSL rehash
command line tool. Fixed in OpenSSL 3.0.4 (Affected 3.0.0,3.0.1,3.0.2,3.0.3).
Fixed in OpenSSL 1.1.1p (Affected 1.1.1-1.1.1o). Fixed in OpenSSL 1.0.2zf
(Affected 1.0.2-1.0.2ze).

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2022-2068

(LOCAL REV: NOT UPSTREAM) -- wrspecific

[OP: keep old openssl 1.1.1o recipe]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>