openssl: upgrade 1.1.1p -> 1.1.1q
Issue: LIN1019-8536

AES OCB mode for 32-bit x86 platforms using the AES-NI assembly optimised
implementation will not encrypt the entirety of the data under some
circumstances. This could reveal sixteen bytes of data that was preexisting in
the memory that wasn't written. In the special case of "in place" encryption,
sixteen bytes of the plaintext would be revealed. Since OpenSSL does not
support OCB based cipher suites for TLS and DTLS, they are both unaffected.
Fixed in OpenSSL 3.0.5 (Affected 3.0.0-3.0.4). Fixed in OpenSSL 1.1.1q
(Affected 1.1.1-1.1.1p).

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2022-2097

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

[OP: keep old openssl 1.1.1p recipe]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>